chromium/third_party/blink/web_tests/http/tests/loading/wbn/wbn-subresource-origin-trial.https.html

<!DOCTYPE html>
<head><title>Subresource Web Bundles Origin Trial</title></head>
<body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
promise_test(async (t) => {
  assert_false(
    document.createElement('link').relList.supports('webbundle'),
    'Subresource Web Bundles should not be supported by default.');

  const meta = document.createElement('meta');
  meta.httpEquiv = "origin-trial";
  // This Origin Trial token is generated with the command:
  //   tools/origin_trials/generate_token.py \
  //     --expire-timestamp=2000000000 \
  //     --version=3 \
  //     https://127.0.0.1:8443 SubresourceWebBundles
  meta.content = "A1nOn5e148yGA6ExfhqzlxQFFC71b03gYEAJEZ1XpUw+Lv8uUA4rpeyhCme0z3bBaRtIy8XkDK8twDUuhVXOegEAAABeeyJvcmlnaW4iOiAiaHR0cHM6Ly8xMjcuMC4wLjE6ODQ0MyIsICJmZWF0dXJlIjogIlN1YnJlc291cmNlV2ViQnVuZGxlcyIsICJleHBpcnkiOiAyMDAwMDAwMDAwfQ==";
  document.head.appendChild(meta);

  assert_true(
    document.createElement('link').relList.supports('webbundle'),
    'Subresource Web Bundles should be supported.');

  const wbn_url = 'https://127.0.0.1:8443/loading/wbn/resources/wbn/wbn-subresource-origin-trial.wbn';
  const script_url = 'https://localhost:8443/loading/wbn/resources/server/wbn-subresource-origin-trial/script.js';

  const link = document.createElement('link');
  link.rel = 'webbundle';
  link.href = wbn_url;
  link.resources.add(script_url);
  document.body.appendChild(link);

  const result_promise = new Promise((resolve) => {
      // This function will be called from script.js
      window.report_result = resolve;
    });

  const script = document.createElement('script');
  script.src = script_url;
  document.body.appendChild(script);

  assert_equals(
    await result_promise,
    'from web bundle',
    'Script should be loaded from the web bundle.');
}, 'Subresource Web Bundles Origin Trial');
</script>
</body>
